The parent of the process may not be the shell
In an EC2 instance with Amazon Linux, dumpcap was running but
is_dumpcap_running said it wasn't. The problem may be that in
CentOS (the base of Amazon Linux) the shell=True does not make the popen
command to run under a shell, but rather directly as a process. So there
is no parent, the Popen object is the dumpcap process itself.