VirtualBox image + different TBB gpg signing key
The provisioning works as it is in VirtualBox too.

hkps://keyserver.cns.vt.edu requires a CA cert to be installed. In
https://www.torproject.org/docs/verifying-signatures.html.en, Tor
Project uses sks-keyservers, which does not require any further step.

Also, the fingerprint they use is different from one we used before: I
think it's a good idea to stick to the one publicly advertised in the Tor
Project website.