Digital Ocean deployment.
Instructions about setting up the provider have been added to
AUTOMATION.md.

skel/digital-ocean-machines provides a template to configure the Digital
Ocean deployment.

vagrant rsync complains about absolute path symlinks, because obviously
in the remote machine they are not going to work. I've changed
pytbcrawler so that results/latest_crawl is a relative path symlink
instead.

We also need to install gcc for a pip package, and libdbus-glib-1-2 for
tbb to work on vanilla Debian 8.

We do not need the tor deb package because TBB already provides it. On
the other hand, using a custom built tor binary is also an interesting
option that we may support soon. I leave the start-tor-daemon.yml file
because we may use it anyways, if we manage to install the service
appropriately.